Move some methods to utils module
