+++ hsolver/utils.py
@@ -1,4 +1,59 @@
 """Utilities module"""
+from typing import List, Dict
+
+import numpy as np
+
+
+def push_axes(tensor: np.ndarray, new_positions: List[int]) -> np.ndarray:
+    """Pushes first axes of the input tensor to specific positions
+
+    :param tensor: Input tensor
+    :param new_positions: New positions of first len(new_positions) axes
+    :return: Resulting tensor
+    """
+    axes = list(range(len(new_positions), tensor.ndim))
+    for idx_position in np.argsort(new_positions):
+        axes.insert(new_positions[idx_position], idx_position)
+    return tensor.transpose(axes)
+
+
+def transform_sv_tensor(
+        state_tensor: np.ndarray, unitary: np.ndarray, transformation_dims: Dict[int, int]
+) -> np.ndarray:
+    """Transforms state-vector tensor with some unitary matrix
+
+    :param state_tensor: Input state-vector tensor
+    :param unitary: Unitary matrix
+    :param transformation_dims: Dimensions in which the matrix is acting
+    :return: Output state-vector tensor
+    """
+    subsystems_indices = list(transformation_dims.keys())
+    subsystems_dims = list(transformation_dims.values())
+    subsystems_count = len(subsystems_indices)
+
+    # transform tensor
+    state_tensor = np.tensordot(
+        unitary.reshape(2 * subsystems_dims),
+        state_tensor,
+        axes=(range(subsystems_count, 2 * subsystems_count), subsystems_indices)
+    )
+
+    # the convoluted axis are put in the beginning, so we need to bring them back
+    state_tensor = push_axes(state_tensor, subsystems_indices)
+
+    return state_tensor
+
+
+def mkron(*matrices) -> np.ndarray:
+    """Performs multiple matrices Kroneker product
+
+    :param matrices: Input matrices
+    :return: Resulting matrix
+    """
+    matrix = np.eye(1)
+    for m in matrices:
+        matrix = np.kron(matrix, m)
+    return matrix
 
 
 class ProgressPrinter:
